Separated utils and custom hooks to own package
