+++ packages/dw-react-controls/package.json
@@ -52,6 +52,7 @@
 		"webpack": "^4.44.2"
 	},
 	"dependencies": {
+		"@dlw-digitalworkplace/dw-react-utils": "^0.1.0",
 		"clsx": "^1.1.1",
 		"rfdc": "^1.3.0",
 		"ts-node": "^9.0.0"
