+++ packages/dw-react-controls/src/components/TaxonomyPicker/TaxonomyPicker.base.tsx
@@ -1,3 +1,4 @@
+import { useStateIfMounted } from "@dlw-digitalworkplace/dw-react-utils";
 import { ValidationState } from "office-ui-fabric-react";
 import { IconButton } from "office-ui-fabric-react/lib/Button";
 import { Label } from "office-ui-fabric-react/lib/Label";
@@ -5,7 +6,6 @@ import { classNamesFunction } from "office-ui-fabric-react/lib/Utilities";
 import * as React from "react";
 import { ITermCreationResult, ITermValue } from "../../models";
 import { ITerm } from "../../models/ITerm";
-import { useStateIfMounted } from "../../utils";
 import { TermPicker } from "../TermPicker";
 import { ITaxonomyPickerProps, ITaxonomyPickerStyleProps, ITaxonomyPickerStyles } from "./TaxonomyPicker.types";
 import { TaxonomyPickerDialog } from "./TaxonomyPickerDialog";
