+++ packages/dw-react-utils/just.config.ts
@@ -1,0 +1,4 @@
+const { preset, just } = require("@dw-react-lib/build");
+const { chain, task } = just;
+
+preset();
