+++ packages/dw-react-utils/package.json
@@ -1,0 +1,35 @@
+{
+	"name": "@dlw-digitalworkplace/dw-react-utils",
+	"version": "0.1.0",
+	"main": "lib-commonjs/index.js",
+	"module": "lib/index.js",
+	"typings": "lib/index.d.ts",
+	"author": "Nick Sevens",
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/dlw-digitalworkplace/dw-react-lib"
+	},
+	"scripts": {
+		"build": "just-scripts build",
+		"clean": "just-scripts clean"
+	},
+	"devDependencies": {
+		"@dw-react-lib/build": "^1.0.0",
+		"@dw-react-lib/tslint-rules": "^1.0.0",
+		"@types/node": "^14.11.8",
+		"@types/react": "^16.9.51",
+		"react": "^16.13.1",
+		"tslint": "^6.1.3",
+		"typescript": "^3.9.7",
+		"webpack": "^4.44.2"
+	},
+	"dependencies": {
+		"clsx": "^1.1.1",
+		"rfdc": "^1.3.0",
+		"ts-node": "^9.0.0"
+	},
+	"peerDependencies": {
+		"react": ">=16.8.0 <17.0.0"
+	}
+}
