+++ packages/dw-react-utils/src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from "./collectionUtils";
 export * from "./useControlled";
 export * from "./useIsMounted";
 export * from "./useStateIfMounted";
