+++ packages/dw-react-utils/src/utils/collections/index.ts
@@ -1,0 +1,2 @@
+export * from "./findInTree";
+export * from "./flatten";
