+++ packages/dw-react-utils/tslint.json
@@ -1,0 +1,7 @@
+{
+	"extends": ["@dw-react-lib/tslint-rules"],
+	"rules": {
+		"jsx-ban-props": false,
+		"no-any": false
+	}
+}
